Splits go lambda into two separate builds with a shared lib. Adds prototype terraform plumbing to deploy both lambdas. Upgrades go to 1.24.
