+++ infra/modules/api_gateway_backend/output.tf
@@ -1,0 +1,3 @@
+output "domain" {
+  value = aws_apigatewayv2_domain_name.api.domain_name_configuration[0].target_domain_name
+}
