+++ infra/modules/api_gateway_backend/variables.tf
@@ -1,0 +1,27 @@
+variable "allowed_cors_origin" {
+  type = string
+}
+
+variable "cert_arn" {
+  type = string
+}
+
+variable "domain" {
+  type = string
+}
+
+variable "lambda_function_album_invoke_arn" {
+  type = string
+}
+
+variable "lambda_function_album_name" {
+  type = string
+}
+
+variable "lambda_function_entries_invoke_arn" {
+  type = string
+}
+
+variable "lambda_function_entries_name" {
+  type = string
+}
