+++ infra/modules/cloudflare_api_dns/variables.tf
@@ -1,0 +1,11 @@
+variable "zone_id" {
+    type = string
+}
+
+variable "domain" {
+    type = string
+}
+
+variable "value" {
+    type = string
+}
