+++ infra/modules/lambda_entries/output.tf
@@ -1,0 +1,7 @@
+output "arn" {
+  value = aws_lambda_function.backend_entries.arn
+}
+
+output "invoke_arn" {
+  value = aws_lambda_function.backend_entries.invoke_arn
+}
