+++ infra/modules/lambda_entries/variables.tf
@@ -1,0 +1,11 @@
+variable "aws_region" {
+  type = string
+}
+
+variable "lambda_deploy_bucket" {
+  type = string
+}
+
+variable "bucket" {
+  type = string
+}
